requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.13
django-allauth==0.51.0
django-braces==1.15.0
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
django-dotenv==1.4.2
django-oauth-toolkit==2.0.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-sendsms==0.5
django-taggit==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-social-oauth2==1.2.1
drfpasswordless==1.5.8
ecdsa==0.17.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.3.1
MarkupSafe==2.1.1
oauthlib==3.2.0
phonenumbers==8.12.49
Pillow==9.1.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-environ==0.4.54
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rest-social-auth==7.0.0
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
